move compute_response from api_utils.Utils to emon_api_core.EmonRequestCore
