+++ emon_tools/async_emon_api.py
@@ -40,6 +40,7 @@ from emon_tools.api_utils import MESSAGE_KEY
 from emon_tools.api_utils import SUCCESS_KEY
 from emon_tools.emon_api_core import InputGetType
 from emon_tools.emon_api_core import RequestType
+from emon_tools.emon_api_core import EmonRequestCore
 from emon_tools.emon_api_core import EmonInputs
 from emon_tools.emon_api_core import EmonFeeds
 from emon_tools.emon_api_core import EmonHelper
@@ -110,7 +111,7 @@ class AsyncEmonRequest:
                 msg,
                 response_data
             )
-            result = Ut.compute_response(
+            result = EmonRequestCore.compute_response(
                 response_data
             )
         else:
