+++ emon_tools/emon_api_core.py
@@ -29,9 +29,11 @@ Security and validation:
 from enum import Enum
 import logging
 from typing import Any, Optional, Dict, Union
-from urllib.parse import quote
-
+from urllib.parse import quote_plus, urljoin
+import simplejson as sj
 from emon_tools.api_utils import Utils as Ut
+from emon_tools.api_utils import MESSAGE_KEY
+from emon_tools.api_utils import SUCCESS_KEY
 
 logging.basicConfig()
 
@@ -475,6 +477,45 @@ class EmonHelper:
         return []
 
 
+class EmonRequestCore:
+    """EmonRequest common helper"""
+    @staticmethod
+    def compute_response(
+        response: Union[dict, list, str, None]
+    ) -> tuple[bool, Union[str, list, dict]]:
+        """
+        Computes and interprets the response from Emoncms.
+
+        :param result: The response from Emoncms.
+        :return: A tuple of success status and message.
+        """
+        result = {SUCCESS_KEY: False, MESSAGE_KEY: "Invalid response"}
+        is_dict = isinstance(response, dict)
+        is_json = is_dict\
+            and SUCCESS_KEY in response\
+            and MESSAGE_KEY in response
+        if is_json:
+            result[SUCCESS_KEY] = bool(response[SUCCESS_KEY])
+            extra = Ut.filter_dict_by_keys(
+                input_data=response,
+                filter_data=[SUCCESS_KEY],
+                filter_in=False
+            )
+            if Ut.is_dict(extra, not_empty=True):
+                result[SUCCESS_KEY] = bool(response[SUCCESS_KEY])
+                del result[MESSAGE_KEY]
+                result.update(extra)
+
+        elif is_dict and SUCCESS_KEY in response and len(response) == 1:
+            result[SUCCESS_KEY] = bool(response[SUCCESS_KEY])
+            result[MESSAGE_KEY] = ''
+
+        elif isinstance(response, (list, dict, str, int, float)):
+            result[SUCCESS_KEY] = True
+            result[MESSAGE_KEY] = response
+        return result
+
+
 class EmonInputs:
     """Emon Inputs Api"""
     @staticmethod
@@ -493,7 +534,7 @@ class EmonInputs:
             Optional[List[Dict[str, Any]]]: A list of input dictionaries
             or None if retrieval fails.
         """
-        path = f"/input/get/{quote(node)}" if node else "/input/get"
+        path = f"/input/get/{quote_plus(node)}" if node else "/input/get"
         return path, None
 
     @staticmethod
@@ -540,7 +581,7 @@ class EmonInputs:
         """
         if not node or not name:
             raise ValueError("Node and name must be non-empty strings.")
-        path = f"/input/get/{quote(node)}/{quote(name)}"
+        path = f"/input/get/{quote_plus(node)}/{quote_plus(name)}"
         return path, None
 
     @staticmethod
